refactor: 명확한 용도 구분을 위해 Station dto의 패키지를 따로 생성
