+++ src/main/java/subway/station/StationController.java
@@ -2,6 +2,8 @@ package subway.station;
 
 import org.springframework.http.ResponseEntity;
 import org.springframework.web.bind.annotation.*;
+import subway.station.dto.StationRequest;
+import subway.station.dto.StationResponse;
 
 import java.net.URI;
 import java.util.List;
